requirements.txt
customtkinter
opencv-python
matplotlib
pytesseract
pandas
numpy
Pillow
